rename module to pants.engine.internals.side_effects
[ci skip-rust]

[ci skip-build-wheels]
